refactor(engine): Phase 4 slice 4 — FX gets its own stage package; RE-split co-locates as stages/re_split/
stages/fx/: FXConverter moved verbatim (engine/fx_converter.py -> 29-line
shim) plus convert_resolved_frames — the five-frame conversion block
extracted from HierarchyResolver.resolve(), which now delegates at the
identical unify -> FX -> enrich seam (load-bearing: LTV / property coverage /
lending-group totals / classifier thresholds assume reporting-currency
amounts). Registry-stage promotion is deferred with a recorded decision —
it needs two new intermediate edge contracts, a dedicated-slice job; the
plan's mid-run config.with_fx_rate hoist already landed in slice 1.

stages/re_split/: splitter.py (engine/re_splitter.py moved verbatim,
producer seal + RE001 keep-alive intact; 23-line shim left), flagging.py
(the classifier's candidate-flagging brain from stages/classify/
re_split_flags.py — classify still invokes it at the same point, only the
home changes), stage.py (the fold adapter). The whole RE-split decision
matrix is now one package; parameters stay in data/tables per check 5.

Two number-changing findings RECORDED, not fixed (decision log): the
Art. 123B currency-mismatch multiplier compares against post-FX currency,
and FX-haircut null polarity disagrees between the SFT and comprehensive
paths.

Verbatim moves (diff-verified); citation snapshot re-homed (2 keys);
REFERENCES_REQUIRED_EXEMPT pruned. Gates: full suite green; parity
byte-identical across all four 10k stress configs; ruff/ty/arch_check green.
