+++ docs/appendix/changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Context-era test surface:** sanctioned builder `tests/fixtures/context.py` (`make_context`), `PipelineContext(` added to the builder-conformance hard lint, fold/registry/context pins in `tests/unit/test_orchestrator_fold.py` + `tests/unit/contracts/test_pipeline_context.py`; the stage-execution tests in `tests/unit/test_pipeline.py` now drive the stage adapters through built contexts.
 - **`engine/hierarchy.py` (3,363 LOC) split into `engine/stages/hierarchy/`** per the mandatory stage anatomy: `graph` (parent/ultimate-parent/facility-graph resolution + the four `cp_lookup_*` seals), `ratings` (dual best-rating inheritance), `facility_undrawn` (synthetic undrawn rows incl. the MOF waterfall; the SA-RW preview stays here until its dedicated slice), `unify` (loans/contingents/facility-undrawn concat + facility metadata), `enrich` (QRRE propagation, rating attach, short-term override, property coverage, LTV, lending group), with a thin `resolver.py` keeping `HierarchyResolver` (verbatim `resolve()` recipe + delegating private methods) and `stage.py` the fold adapter. `engine/hierarchy.py` survives as a 28-line back-compat shim, so the 23+ test files importing `HierarchyResolver` from it are untouched. **Function bodies moved verbatim** — the defensive-surface ratchet metrics are byte-identical (fill_null 446, presence guards 374, collect_schema probes 166) and the parity gate stays byte-identical; `max_engine_module_loc` banks 3,364 → 2,252.
 - **`engine/classifier.py` (2,227 LOC) split into `engine/stages/classify/`** per the mandatory stage anatomy: `attributes` (counterparty/SL joins, independent flags, shared SME size-test expr — keeps the `_pt_upper`/`_sa_class` scratch-column builders co-located with `derive_independent_flags`), `subtypes` (SME/retail/QRRE class mutation, Art. 147(5) reclassification, IRB-class sync, B31 subclass), `re_split_flags` (the 6-function RE loan-split candidate block + `_SECURED_TARGET_*`, isolated so the Slice-4 re_split co-location never moves it again), `permissions` (model-permission resolution, permission exprs, CLS006 diagnostics), `approach` (decision ladder + B31 Art. 147A restrictions + `_B31_SLOTTING_ONLY_SL_TYPES`), `audit` (audit trail, CLS008/DQ008 warnings), with a thin `classifier.py` keeping `ExposureClassifier` (verbatim `classify()` recipe — materialise-before-diagnostics, raw `seal()` after, CCR brand probe — plus `_build_bundle`) and `stage.py` the fold adapter. `engine/classifier.py` survives as a 24-line back-compat shim, so the 30 test files importing `ExposureClassifier` from it are untouched; the stale `ENTITY_TYPE_TO_*` re-export comment is deleted (all consumers import from `data.tables.entity_class_mapping` directly). **Function bodies moved verbatim** — ratchet metrics unchanged; the 8 `@cites` decorators moved with their functions and the citation snapshot/matrix re-keyed to the new module paths.
+- **FX conversion and the RE-split each get their stage package (Slice 4).** `engine/stages/fx/` lands the FX code seam: `converter.py` (the stateless `FXConverter` five-method kernel + `create_fx_converter`, moved verbatim from `engine/fx_converter.py`) and `conversion.py` (`convert_resolved_frames` — the five-converter block extracted verbatim from `HierarchyResolver.resolve`, still invoked at the same unify → FX → enrich seam because LTV / property-coverage / lending-group totals and the classifier's GBP thresholds assume reporting-currency amounts). Registry promotion of FX to a standalone stage is deferred — the code seam landed here, the EUR/GBP config-mutation hoist landed in Slice 1; `engine/fx_rate_sync.py` stays with the pipeline facade. `engine/stages/re_split/` co-locates the whole RE loan-split: `splitter.py` (`RealEstateSplitter` + the 19 split/allocation helpers moved verbatim from `engine/re_splitter.py`, producer seal and RE001 keep-alive included), `flagging.py` (the candidate-flagging brain moved verbatim from `stages/classify/re_split_flags.py` — still invoked from `classify()` at the same point), and `stage.py` (the Slice-1 fold adapter, previously `stages/re_split.py`). Split parameters stay in the data layer (`data/tables/re_split_parameters.py`, arch_check check 5). `engine/fx_converter.py` and `engine/re_splitter.py` survive as thin back-compat shims, so the test files importing `FXConverter` / `RealEstateSplitter` from them are untouched. **Function bodies moved verbatim** — no behaviour change (the recorded null-currency FX-haircut findings and the Art. 123B post-FX currency comparison are deliberately left as-is); ratchet metrics unchanged; the 2 moved `@cites` functions re-keyed in the citation snapshot/matrix.
 
 ### Changed (architecture — migration Phase 1: eager stage edges)
 - **Stages now exchange materialised frames; laziness is strictly intra-stage.** `engine/materialise.py` is rewritten around `materialise_edge(lf, config, label)`, called at every stage exit (`hierarchy_exit`, `ccr_exit` when a derivatives book is present, `classifier_exit`, `crm_exit`, `re_split_exit`, and the three calculator branches). The hand-placed barrier inventory — `classifier_output`, `pipeline_pre_branch`, `crm_post_ead_unified`/`_fanout`, `crm_no_guarantee` — is deleted; Two benchmark-justified **intra-stage checkpoints** survive inside CRM: `crm_post_ead` (a controlled A/B showed removing it costs 35–52% on the full-pipeline benchmarks — the collateral lookups re-execute the provisions→CCF→EAD chain without it) and `crm_pre_guarantee_unified` (empirically irreducible on Polars 1.37). Bundle fields stay `LazyFrame`-typed (cheap `.lazy()` wrap) until the Phase 3 producer seal. Byte-identical outputs; the inter-stage plan-depth SIGSEGV class is now unrepresentable. See the rewritten [Stage-Edge Materialisation](../architecture/pipeline-collect-barriers.md) page.
