+++ docs/plans/target-architecture-migration.md
@@ -516,4 +516,7 @@ annotations.
 | 2026-06-12 | Phase 4 slice 1: components are built **per run** from the effective config; the orchestrator caches nothing across runs | Deliberate behaviour change (improvement) over the pre-fold lazy-init cache: `CRMProcessor(is_basel_3_1=…)` constructor state made a reused orchestrator framework-stale (the documented comparison.py two-orchestrator workaround). Construction cost is negligible (the haircut table is a small eager frame). The workaround in comparison.py is now redundant but harmless — removed when CRM constructor-state dies (Phase 5) |
 | 2026-06-12 | Phase 4 slice 1: adapter-era context grain is **whole bundles** (RAW_DATA, RESOLVED_HIERARCHY, CLASSIFIED, CRM_ADJUSTED, …), not per-frame keys; the four error channels keep their exact pre-fold merge order, incl. the calc/aggregator-failure double-conversion quirk | Bundle `__post_init__` brand validation keeps firing free at every stage boundary, and zero behaviour change is provable by the parity gate. Per-frame keys land slice-by-slice as stages convert to the uniform anatomy; the error channels collapse in the dedicated slice (P2.21) where the observable code changes are reviewed once, deliberately |
 
+| 2026-06-12 | Phase 4 slice 4: FX conversion gets its own package (`stages/fx/` — FXConverter + `convert_resolved_frames`, resolver delegates at the identical unify→FX→enrich seam); **registry-stage promotion deferred** with this recorded decision. RE-split co-located as `stages/re_split/` (splitter + the classifier's flagging brain + stage adapter; shims left behind) | The FX seam is intra-hierarchy and load-bearing (LTV / property coverage / lending-group totals / classifier GBP thresholds all assume reporting-currency amounts); promoting FX to a registered stage requires two new intermediate edge contracts (pre-FX hierarchy state) whose parity verification belongs to a dedicated slice, not mid-move. The plan's other FX requirement — the mid-run `config.with_fx_rate` mutation moving out of the fold — landed in slice 1 (facade hoists eur_gbp sync before components/rulepack are built) |
+| 2026-06-12 | Phase 4 findings (recorded, NOT fixed — number-changing): (a) `sa/namespace.py apply_currency_mismatch_multiplier` (Art. 123B / PS1/26) tests borrower income currency against the post-FX `currency` column (= base currency on every converted row), not `denomination_currency_expr`/`original_currency` — a USD loan to a USD-income borrower is flagged mismatched after conversion to GBP; (b) FX-haircut null polarity disagrees across paths: `ccr/sft_fccm.py` applies H_fx when either currency is null (conservative) while `crm/haircuts.py`/`crm/guarantees.py` skip it (anti-conservative, the Phase 3 recorded `original_currency`-null finding) | Both change B3.1/CRM numbers; per the do-not-do register they need fixture verification + oracle coverage + a recorded preserve-or-fix decision each, not a silent flip during a move slice. (a) is P1.135-class — verify with a fixture before the namespace-retirement slice moves that code; (b) unify polarity in the CRM/kernel slice with a parity-gate decision |
+
 *(Append further preserve-or-fix decisions here as phases land.)*
